project/delivery_2/spotify_like/requirements.txt
Arpeggio==2.0.2
PyYAML==6.0.2
textX==4.1.0
